File refactoring
