+++ frontend/package.json
@@ -9,8 +9,8 @@
     "build": "npm run build:css && tsc -b && vite build",
     "build:prod": "npm run build:css && tsc -p tsconfig.json && vite build --mode production",
     "build:css": "npm run build:css-light && npm run build:css-dark",
-    "build:css-light": "tailwindcss -i ./src/main.light.css -o ./public/main.light.css --mode=production",
-    "build:css-dark": "tailwindcss -i ./src/main.dark.css -o ./public/main.dark.css --mode=production",
+    "build:css-light": "tailwindcss -i ./src/common/styles/main.light.css -o ./public/main.light.css --mode=production",
+    "build:css-dark": "tailwindcss -i ./src/common/styles/main.dark.css -o ./public/main.dark.css --mode=production",
     "lint": "eslint .",
     "preview": "vite preview",
     "storybook": "storybook dev -p 6006",
