+++ frontend/src/App.tsx
@@ -1,13 +1,15 @@
 import { StrictMode } from 'react';
-import { Provider } from 'react-redux';
-import { appStore } from './store';
 import { createRoot } from 'react-dom/client';
+import { Provider } from 'react-redux';
 import { BrowserRouter } from 'react-router-dom';
+
 import '@radix-ui/themes/styles.css';
-import * as Tooltip from '@radix-ui/react-tooltip';
 import { Theme } from '@radix-ui/themes';
+import * as Tooltip from '@radix-ui/react-tooltip';
+
+import { ThemeProvider, useTheme } from './common/contexts/ThemeContext';
+import { appStore } from './common/store';
 import AppRoutes from './AppRoutes.tsx';
-import { ThemeProvider, useTheme } from './contexts/ThemeContext';
 
 const defaultAppearance = localStorage.getItem('theme') || 'dark';
 import(`./main.${defaultAppearance}.css`);
