+++ frontend/src/common/stories/base/Background.stories.tsx
