+++ frontend/src/common/styles/main.dark.css
@@ -4,15 +4,12 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-@config "./../tailwind.config.js"; 
-
+@config "./../../../tailwind.dark.config.js"; 
 
 body {
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
-
-
 /* —————————————————————————————————————————————————————————————————————————————
   Fonts
   https://fonts.google.com/selection/embed
