+++ frontend/src/common/styles/main.light.css
@@ -4,12 +4,15 @@
 @tailwind base;
 @tailwind components;
 @tailwind utilities;
-@config "./../tailwind.dark.config.js"; 
+@config "./../../../tailwind.config.js"; 
+
 
 body {
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
+
+
 /* —————————————————————————————————————————————————————————————————————————————
   Fonts
   https://fonts.google.com/selection/embed
