+++ frontend/src/components/Debug.tsx
@@ -6,7 +6,7 @@ import { Tooltip } from '@components/Tooltip';
 
 import { Menu, MenuButton, MenuItem, MenuItems } from '@headlessui/react';
 import { ToolbarButtonClassName } from '@/module/book/ui/ToolbarButton';
-import { useTheme } from '@/contexts/ThemeContext';
+import { useTheme } from '@/common/contexts/ThemeContext';
 
 type ThemeAppearance = 'light' | 'dark';
