+++ frontend/src/module/book/book.actions.ts
@@ -4,7 +4,7 @@ import { BookService } from '@/services/book.service';
 import { Book, Page, Element } from '@/types/book';
 import { ElementService } from '@/services/element.service';
 import { GraphicAsset } from '@/types/graphic-asset.entity';
-import { PageService } from '@/services/page.service';
+// import { PageService } from '@/services/page.service';
 
 /*––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––––
  * BOOKS_FETCH_BOOK_BY_ID
