+++ frontend/src/module/book/canvas/element/DrawableElement.ts
@@ -1,6 +1,6 @@
 import * as fabric from 'fabric';
 import { Element } from '@/types/book';
-export interface DrawableObject {
+export interface DrawableElement {
   draw(canvas: fabric.Canvas): void;
   update(attrs: Element): void;
   delete(): void;
