+++ frontend/src/module/book/ui/BookHeader.tsx
@@ -1,7 +1,7 @@
 import BreadCrumb from '@/components/BreadCrumb';
 import Header from '@/components/Header';
 import InlineEdit from '@/components/InlineEdit';
-import { useAppSelector } from '@/hooks/useRedux';
+import { useAppSelector } from '@/common/hooks/useRedux';
 import { Book } from '@/types/book';
 import { selectIsLoading } from '../book.state';
