+++ frontend/src/module/home/DescriptionSection.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import { motion } from 'motion/react';
+
 import AnimatedText from '@/components/AnimatedText';
 import Button from '@components/Button';
 import {
   QueueListIcon,
-  // BookOpenIcon,
   PlusIcon,
 } from '@heroicons/react/24/outline';
 import { Tooltip } from '@/components/Tooltip';
